refactor: 모듈화하여 gear 라우트 파일 길이 최적화
server/routes/gear.ts를 469줄에서 30줄로 단축:
- 기본 CRUD 작업을 basicOperations.ts로 분리
- PATCH 부분 업데이트를 patchOperations.ts로 분리
- 벌크 작업을 bulkOperations.ts로 분리
- 히스토리 관리를 historyOperations.ts로 분리
